object-recognition-service/requirements.txt
ultralytics>=8.0.0
numpy>=1.22.0
Pillow>=9.0.0
pika>=1.3.2
minio>=7.2.15